clean up utility and servicer module
